Cleaned up GameObject code a bit.
Didn't make any radical changes. I just wanted to get an
understanding of how they worked.

There are some funky things in it. Like how b2body
doesn't exist in GameObject but it has a method in preperatio nfor it.

Eyes are getting tired, so I'll look at it again tomorrow and see if I can
improve it or not.