Added alternate controls.
Added a new set of controls - arrow keys.
Modified KeyboardManager & Keyboard - a new enum called Action is now used
to get key inputs. For example, checking for Action::MOVE_LEFT checks for
Key::A OR Key::LEFT.

This interface is more semantic and allows for multiple control schemes.
It'll also be slightly easier to transition to user-editable controls,
although we may never reach this implementation.